feat(quality): enforce spotbugs gate with zero findings
Enable the SpotBugs plugin (report mode first) and clear all 117 findings:
- rename the two ReconciliationException data classes (named like Java
  exceptions but are reconciliation records): entity ->
  ReconciliationExceptionRecord, dto -> ReconciliationExceptionItem
- BusinessMetrics registers metrics in @PostConstruct instead of the
  constructor (overridable call from constructor)
- fix the real NPE path in BillingPlanServiceImpl.saveSegments (null
  segments NPEs on the for-loop despite the guard only wrapping
  validation) and remove a useless PaymentResponse allocation
- verifyWechatSignature returns Optional<Boolean>, unbox/rebox
  eliminations, null-safe session URI/remote-address extraction,
  equals symmetry (callSuper), switch default
- documented exclusions for EI_EXPOSE_REP/EI_EXPOSE_REP2 (Lombok data
  classes by design) and REC_CATCH_EXCEPTION (third-party boundary)

With findings at zero the gate flips to fail-on-finding.

+++ evcs-order/src/main/java/com/evcs/order/dto/stats/ChargerActiveDays.java
@@ -4,6 +4,6 @@ package com.evcs.order.dto.stats;
  * 仪表盘统计投影（内部 API 返回，字段与归属服务数据对齐）。
  */
 public record ChargerActiveDays(
-        Long chargerId, Integer activeDays
+        Long chargerId, Long activeDays
 ) {
 }
