+++ evcs-order/src/main/java/com/evcs/order/dto/stats/HourlyCount.java
@@ -4,6 +4,6 @@ package com.evcs.order.dto.stats;
  * 仪表盘统计投影（内部 API 返回，字段与归属服务数据对齐）。
  */
 public record HourlyCount(
-        Integer hour, Long count
+        Long hour, Long count
 ) {
 }
